Revert "Vue router assignment"
