+++ src/components/UI/BaseCard.vue
@@ -1,0 +1,19 @@
+<script setup>
+
+</script>
+
+<template>
+    <div>
+        <slot></slot>
+    </div>
+</template>
+
+<style scoped>
+div{
+    border: 1px solid rgb(86, 87, 87)21;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+    border-radius: 10px;
+    padding: 25px;
+    margin: 10px
+}
+</style>
