+++ src/main.js
@@ -1,0 +1,19 @@
+import { createApp } from 'vue'
+import './style.css'
+import App from './App.vue'
+import { library } from '@fortawesome/fontawesome-svg-core'
+import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
+import { faSort } from '@fortawesome/free-solid-svg-icons'
+
+
+library.add(faSort)
+
+
+
+
+const app = createApp(App)
+
+// app.component('font-awesome-icon', FontAwesomeIcon)
+app.component('font-awesome-icon', FontAwesomeIcon)
+
+app.mount('#app')
