+++ src/style.css
@@ -1,5 +1,5 @@
 :root {
-  font-family: sans-serif;
+  font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
   line-height: 1.5;
   font-weight: 400;
 
@@ -12,15 +12,7 @@
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 }
-#card {
-  border: 1px solid rgb(205, 207, 207) 21;
-  box-shadow: 0 4px 8px #6fb3b8;
-  background-color: #f6f6f2 ;
-  border-radius: 10px;
-  padding: 25px;
-  margin: 10px;
-  
-}
+
 a {
   font-weight: 500;
   color: #646cff;
@@ -36,14 +28,11 @@ body {
   place-items: center;
   min-width: 320px;
   min-height: 100vh;
-  background-color: #f6f6f2;
-  color: #388087;
 }
 
 h1 {
   font-size: 3.2em;
   line-height: 1.1;
-  margin: 2px;
 }
 
 button {
@@ -52,11 +41,10 @@ button {
   padding: 0.6em 1.2em;
   font-size: 1em;
   font-weight: 500;
-  font-family: sans-serif;
-  background-color: #badfe7;
+  font-family: inherit;
+  background-color: #1a1a1a;
   cursor: pointer;
   transition: border-color 0.25s;
-  margin: 5px;
 }
 button:hover {
   border-color: #646cff;
@@ -66,99 +54,18 @@ button:focus-visible {
   outline: 4px auto -webkit-focus-ring-color;
 }
 
+.card {
+  padding: 2em;
+}
+
 #app {
   max-width: 1280px;
   margin: 0 auto;
+  padding: 2rem;
   text-align: center;
 }
-td {
-  border-bottom: 1px solid #6fb3b8;
-  padding: 8px;
-  background-color: #f6f6f2;
-}
 
-td #task-name{
-  font-size: 1em;
-  text-align: left;
-}
-
-th {
-  border-radius: 10px 10px 0 0;
-  border-bottom: 1px solid #6fb3b8;
-  padding: 8px;
-  background-color: #388087;
-  font-size: 1.2em;
-}
-
-th {
-  padding-top: 12px;
-  padding-bottom: 12px;
-  text-align: center;
-  color: white;
-}
-#title:hover {
-  background-color: #4bacb4;
-  cursor: pointer;
-}
-#status:hover {
-  background-color: #cbcbc7;
-  cursor: pointer;
-}
-.active{
-  color: green
-}
-.notActive{
-  color:red
-}
-#actions {
-  margin: 5px;
-}
-#main {
-  display: flex;
-}
-#check-box{
-  display: flex;
-  justify-content: center;
-}
-#filter {
-  display: inline-block;
-  border: 1px solid rgb(205, 207, 207) 21;
-  box-shadow: 0 4px 8px #6fb3b8;
-  background-color: #f6f6f2 ;
-  border-radius: 10px;
-  padding: 25px;
-  margin: 10px;
-  height: 200px;
-}
-#filter div {
-  text-align: left;
-}
-#filter section{
-  margin-bottom: 10px;
-}
-
-#loader h3{
-  margin: 2px;
-}
-dialog{
-  font-size: 2rem;
-  color: #388087;
-  border: 1px solid white;
-  box-shadow: 0 4px 8px #6fb3b8;
-  background-color: #f6f6f2 ;
-  border-radius: 10px;
-
-}
-#editable-task:focus
-#editable-task:focus-visible{
-  border-color: white;
-}
-#links{
-  margin: 15px;
-  font-size: 1.2rem;
-}
-
-/* @media (prefers-color-scheme: light) {
+@media (prefers-color-scheme: light) {
   :root {
     color: #213547;
     background-color: #ffffff;
@@ -169,4 +76,4 @@ dialog{
   button {
     background-color: #f9f9f9;
   }
-} */
+}
